[PWA/Trusted Icons] Skip icon reading logs for trusted icons
The WebAppIconManager logs a lot of information in the terminal whenever
it comes across errors while reading icons. For trusted icons reading,
oftentimes the behavior falls back to reading manifest icons from the
disk if not found, leading to too many logs in the terminal to the
point that they are spammy. This CL fixes that by skipping logs which
arise when trusted icons are not found on the disk, as the fallback
option will still work or log if there is any error.

Bug: 427566193
Change-Id: I9e0478585e178b604f6f911936aaefaebf9e1c8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6896857
Reviewed-by: May Siem <msiem@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508353}