[CrossDevicePrefTracker] Introduce CrossDevicePrefTrackerFactory
This change introduces CrossDevicePrefTrackerFactory, a
ProfileKeyedServiceFactory responsible for creating and managing
CrossDevicePrefTracker instances for chrome/browser.

This is the first step in making the service available to clients.
Future CLs will expose CrossDevicePrefTracker to Chrome on Android and
Chrome on iOS.

Change-Id: I287c6c0b7fe6b52920b133b9262c324e8cd20a8f
Fixed: 441819928
Low-Coverage-Reason: OTHER
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6898334
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Cr-Commit-Position: refs/heads/main@{#1508377}