Add support for requestFullscreen to a secondary display on Android
This commit introduces the usage of the "screen" option of the
Element.requestFullscreen() function for Chrome on Android. The
blink::mojom::FullscreenOptions::displayId is now populated through the
delegates layer to the FullscreenManager.

The behavior of entering the full screen is coherent with the desktop
implementations. The requesting window is moved to the target display
and a full screen is triggered there. When exiting the full screen the
window is moved back to the originating position.

For window move the Task.moveTaskTo was used, it requires Android API
level 36 and Chrome being set as the default browser.

Moving the window to another display may trigger Activity recreation. To
keep the full screen enabled after the recreation the tab restore
procedure was modified. When tab is recreated the new full screen
TabAttributes are checked to verify the full screen expected state.

Feature flag: EnableFullscreenToAnyScreenAndroid

Screencast:
http://screencast/cast/NTM3MTU4ODIzMTgyMzM2MHwyODRmZjZlYS00Ng

Bug: 373304413
Change-Id: Iab6c0256ca0afa2bc569b7910769e3765c2f0dd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6446897
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Auto-Submit: Wojciech Zwonarz <wzwonarz@google.com>
Commit-Queue: Wojciech Zwonarz <wzwonarz@google.com>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508378}