Fix bug where ColorPrevalence was not sufficiently respected.
If the registry had a valid accent color but ColorPrevalence was 0, the
Windows native color mixer would avoid overwriting the various sys
colors, but the NativeTheme's user_color would still be set to a
non-null value, which would be written into the ColorProviderKey. If the
"follow device theme" preference was on (and thus the color provider
source was changed from "baseline" to "accent" in browser_frame.cc),
this would result in the code in ref_color_providers.cc behaving as if
there was a valid accent color, resulting in a strange and untested
state.

The root cause of this was really that the AccentColorObserver API was
sharp-edged: callers were supposed to ignore the value of accent_color()
etc. when use_dwm_frame_color() was false. Since the native color mixer
code didn't do that when updating the NativeTheme, we got the bug above
(and didn't notice when testing, since the test cleared the "use DWM
frame color" bit when clearing the accent color).

Fix by simplifying the AccentColorObserver API to not have that flag at
all: either there is an accent color (and maybe an inactive and/or
border color), or there isn't.

This also cleans up the relevant test a bit.

Bug: 431953872
Change-Id: I4e6def1679ed257af891c80fad842ae0a469bf9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6897790
Reviewed-by: Greg Thompson <grt@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Tom Lukaszewicz <tluk@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508398}