Revert PasswordFormManagerObserver on IOS
This reverts the `PasswordFormManagerObserver` related code on IOS to
the state prior to
https://chromium-review.googlesource.com/c/chromium/src/+/6703663.

The changes ended up causing crashes (crbug.com/441585012 ). So far we
were not able to reproduce the crashes, so we need to revert as the
crash is release-blocking for stable.

BYPASS_RECITATION_REASON=reverting code

Bug: 441585012
Change-Id: I0ae7c236ba3d33f4612ed4103f847f9efdcf862a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6898279
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Oleksandr Tara <otara@google.com>
Cr-Commit-Position: refs/heads/main@{#1508417}