Refactor incognito check in tab reorder prompt and update lastId in TabModelImpl.
In `SourceViewDragDropReorderStrategy`, the incognito check in
`shouldShowUserPrompt` is moved to the start of the method to ensure no
prompt is shown for incognito tabs.

In `TabModelImpl.setIndex`, the `lastId` is now obtained from the
`mCurrentTabSupplier` instead of calling `getLastId(type)`.

Bug: 441662150
Change-Id: I01b35e55226cb2dd2fa241805c362f40de59ab18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6899869
Auto-Submit: Hitarth Kothari <hitarthkothari@google.com>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Hitarth Kothari <hitarthkothari@google.com>
Cr-Commit-Position: refs/heads/main@{#1508451}