Fix presubmit string concatenation for multiline explanations
Some of the multiline explanations are missing a trailing space so
when the explanation is being used when trying to upload a CL, some of
the words are connected together without a space separating them. This
CL adds a trailing space to some of the multiline explanations.

Bug: none
Change-Id: I127c97ec731ecb8a81b84af8df3d140dc688cb6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6896721
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Steven Luong <stluong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508458}