[iOSImportPasswordsSafari] Added string for policy-blocked items
This CL adds a new string to inform the user when a specific data type
cannot be imported from Safari because it is disabled by an enterprise
policy.

The string will be used in an upcoming CL that updates the Safari Data
Importer UI logic.

Bug: 407587751

Change-Id: Icdadebd9a1ff054d8d56bd7968b74b7744d485d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6896752
Commit-Queue: Sourav Uttam Sinha <sinhasourav@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Alexis Hétu <sugoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508459}