[AriaNotify] On Mac, fire on the node rather than the associated window
This was discussed with James Craig at Apple here:
https://github.com/w3c/aria/pull/2577#discussion_r2305396974

Today, there is no functional difference with this change, but there
could be in the future, especially considering that the element/node
this is called on is intended to inform ATs about the language the
announcement should be read in.

To facilitate, `PostAnnouncementNotification` was refactored to take an
`id` instead of a NSWindow*, making it closer match the API it calls,
`NSAccessibilityPostNotificationWithUserInfo`. This part is pure
refactoring, and will not have any behavioral change.

Change-Id: I1c31273338602d420c6858decd680c295f2eb7ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6897470
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Jacques Newman <janewman@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1508466}