Glic multi instance: Remove Host from GlicKeyedService
GlicWindowControllerImpl now holds the Host when the GlicMultiInstance
flag is off. A host() method is temporarily added to
GlicWindowController to make this accessible during the refactor, so
GlicPanelCoordinatorImpl must also provide a host() for now.

This is to be the first of several CLs that remove things from
GlicKeyedService that should eventually belong to instances.

Bug: b:441424932
Change-Id: I4efe7c1e4bc8ddf0c654826caad82cce81226180
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6896656
Commit-Queue: Ian Wells <iwells@chromium.org>
Reviewed-by: Trevor Perrier <perrier@chromium.org>
Auto-Submit: Ian Wells <iwells@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508475}