[NTPShortcuts] Update MostVisitedSites to support enterprise shortcuts
This change adds enterprise shortcuts functionality to MostVisitedSites.
This change does not include frontend changes (TS/handlers) to view
the enterprise shortcuts. If a user selects the enterprise shortcuts
view in the NTP, top sites will be shown instead. A follow-up change
will include the frontend changes

To achieve that, this change does the following:
- Create EnableTileTypesOptions for MostVisitedSites::EnableTileTypes
to take as option to determine what tiles to enable
- Update MostVisitedSites to call EnterpriseShortcutsManager to fetch
shortcuts and trigger user updates, reorders, and deletions
- Update MostVisitedSites to convert enterprise shortcuts to NTPTiles
which are sent to the NTP frontend
- Update the NTPTiles struct to include "allow_user_edit" and
"allow_user_delete" booleans which will be used to determine action
menu disablement in the frontend (by default true and only set for
enterprise shortcuts)

This change has no UI functionality changes.

Bug: b:438304256
Change-Id: I25365ce148a870bdc3012734e65cd29bb90f6b04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6897126
Commit-Queue: Alex Chen <alexwchen@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508509}