[crd host][wayland] Capture cursor image on second monitor
While the cursor position can be captured on any stream, the cursor
image can only be capture on the stream where the cursor is currently
located, which means the cursor is not going to change on the second
monitor.

This CL fixes this by trying to capture the cursor from all streams and
reporting the one that is successfully captured.

Bug: 432217140
Change-Id: I936d6b75cc8fff64a0243ca503c1cc95cce5fc8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6897202
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Auto-Submit: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508516}