Do not allow the top bar to wrap
This is an improvement to the logcat filtering site.
This CL makes it so that the top bar no longer wraps and consumes
two rows of space. Instead, when the screen is not wide enough,
there is a horizontal scrollbar underneath the top bar and the
top bar always consumes 1 row of space regardless of the screen size.

Bug: 438720787
Change-Id: I72f652e9d2b6e527056544e3d1a811b9910f2cdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6900057
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Auto-Submit: Martin Kong <martinkong@google.com>
Cr-Commit-Position: refs/heads/main@{#1508518}