Split tab tabstrip announcements
Announce whether we are in the left or right view in a split when
tabbing through the tab strip. To do so, this CL makes Tab update its
accessible name when either the its split id changes, or when the split
contents are rearranged (because split ids don't change in that case).

Translation screenshots:
https://screenshot.googleplex.com/3eNC5BTGUXf5RYf.png
https://screenshot.googleplex.com/PcyEu36WciPuJgu.png

Change-Id: Ib24c8daa42f3e8c2f0cb0be5d3f9a59e96c4d467
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6892230
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Commit-Queue: Charles Meng <charlesmeng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508524}