[NTPShortcuts] Enable enterprise shortcuts feature by default
Enables the NTPEnterpriseShortcuts feature flag by default. This enables
the policy handler to read the NTPShortcuts policy and set the pref. If
no policy is set, the enterprise shortcuts option does not appear on
the NTP and no enterprise shortcuts code is executed.

Follow-up work includes setting the enterprise shortcuts option as
default when the policy is set, IPH for discoverability, and success
metrics.

Enabled-by-default-reason: killswitch

Bug: b:439823163
Change-Id: I0dd8e9302d70882583846c23ebebeede093cfbb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6896338
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Alex Chen <alexwchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508543}