[ntp-composebox] Attach all ContextualInputs to upload payload for non-image requests
The query controller previously only uploaded the first ContextualInput
for PDFs (to match the old functionality). This change makes it so that
all ContextualInputs are attached similarly to LensOverlay.

Replaces CreatePDFFileUploadPayload with
CreateContentextualDataUploadPayload, which accepts page title and url
as well as the data, and compressed pdf data (outside of Android). This
change should enable support for PDF tab context.

Bug: 441142997
Change-Id: I2b060f6c85fd0f6b07ed3d7d481b509602268fee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6892437
Reviewed-by: Nihar Majmudar <niharm@google.com>
Reviewed-by: Duncan Mercer <mercerd@google.com>
Commit-Queue: Jason Hu <hujasonx@google.com>
Cr-Commit-Position: refs/heads/main@{#1508580}