[BTM] Decompose `UrlAndSourceId` type into two fields
This refactors the `UrlAndSourceId` by replacing it with its two fields:
a `GURL` and a `ukm::SourceId`. The `UrlAndSourceId` type originally
replaced the `GURL` type in some fields. However, we only use this type
for labelling nodes in a redirect chain (redirectors, initial referrers,
and final destinations). The `UrlAndSourceId` struct was originally
introduced to enforce that a URL and its associated source ID were
always updated together, but that is no longer a necessary constraint.
Additionally, all uses of the struct already access its individual `url`
or `source_id` fields, making the type unnecessary.

Bug: 433300379
Change-Id: I0e5af5f16b72436787565e2a4a0138cfa55f1103
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6892952
Reviewed-by: Svend Larsen <svend@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Tiffany Song <tiffanysong@google.com>
Cr-Commit-Position: refs/heads/main@{#1508581}