Ensure SyncToken generation in ProduceCanvasResource()
This CL ensures that a SyncToken is always generated after the
WillDrawInternal() call (indirectly invoked via WillDraw()) in
ProduceCanvasResource().

This is part of a prerequisite refactoring to ensure SyncToken
generation in CanvasResourceProvider is always localized and easy
to reason about, so that later we can safely apply scoped SyncToken
access.

Bug: 40286368
Change-Id: If3e7f5d73c59d0489d4e9766c572ad9f2c5c314e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6892956
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Mingjing Zhang <mjzhang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508679}