Better track when WebPreferences need updates for color-related changes.
Currently WebContentsImpl carries a bunch of members it never actually
uses except to determine if the preferences need updating, because it's
doing so on behalf of the ChromeContentBrowserClient. Move this
knowledge there to fix the layering violation.

It also has somewhat-mangled handling of color-related change
notifications that can result in preference changes being sent both too
frequently (e.g. sending web preference updates twice if a color change
triggers both OnNativeThemeUpdated() and OnColorProviderChanged()) and
too infrequently (IIRC, if the color maps change not because of a color
provider change but because of a native theme change; I think this can
happen with system accent color changes?). Simplify this by routing both
types of notifications through a single update function that makes sure
to update everything color-related but only does so when anything has
actually changed.

Bug: 431953872
Change-Id: I897598020881f7e53997f2006dacda2f76efcb9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6886797
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508706}