[PDF] Fix a locale-dependent test
This message was introduced originally in https://crrev.com/c/2610627
and uses locale-dependent serialization for base::Time objects. This was
the intention, since the purpose of this string is for display to the
user not for parsing.

Specifically the OnDocumentLoadComplete test was failing because the
expected metadata message didn't match if the locale or timezone was set
differently.

Change-Id: I06a32c00b171a5c8b4c6b29972b0b84c815531a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6896855
Reviewed-by: Andy Phan <andyphan@chromium.org>
Commit-Queue: April Kallmeyer <ask@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508708}