[AnchoredPopup] Create Builder for AnchoredPopupWindow
This is part of the effort to make AnchoredPopupWindow created through a
builder. The goal is to eliminate as many setters for this class as
possible, and make the popup lifecycle easier to reason about.

To minimize impact, this CL just forwards the method calls from the
builder to the setter, and mark the current setters deprecated.

Bug: 439677294
Change-Id: Ib6da864c79d59e8082cd0acdb1bb5b27b6775e16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6872645
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Reviewed-by: Sinan Sahin <sinansahin@google.com>
Cr-Commit-Position: refs/heads/main@{#1508814}