Fix MockContentBrowserClient in ServiceWorkerBrowserTest
To test synthetic response in `ServiceWorkerBrowserTest`, we use
`MockContentBrowserClient` so that the navigation request is always
eligible for the synthetic response for tests. It used to be OK since
the feature flag was enabled only in
`ServiceWorkerSyntheticResponseBrowserTest`. However, after adding the
entry to the fieldtrial config, the feature is enabled by default in
tests. That leads test failures if other tests in
`ServiceWorkerBrowserTest` uses `MockContentBrowserClient`.

Specifically, following tests are failed.

- ServiceWorkerBrowserTest.FetchPageWithSaveData
- ServiceWorkerBrowserTest.CrossOriginFetchWithSaveData

This CL fixes them by introducing the flag and explicitly set it at the
top of each test.

Bug: 352578800
Change-Id: I239fc73d712b442b9eecf32a01bf455d4b327421
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6893449
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Keita Suzuki <suzukikeita@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508975}