Add more condition to inject service worker for synthetic response
Synthetic response injects service worker if eligible, and dispatches
the network request. Injecting SW is done in `ServiceWorkerRegistry`,
and dispatcing the network request is done in
`ServiceWorkerMainResourceLoader`.

The network request by synthetic response is not dispatched when the
navigation is reloading.  However, the service worker script is already
injected and that incolves service worker interceptor in the navigation
path, which prevents search prefetch. So the condition was not enough to
enable synthetic response on the top of search prefetch.

This CL prevents the browser from injecting the service worker script if
the request is started from reloading, or non-GET request, rather than
disabling the network request from the synthetic response. This change
solves following tests:

- SearchPrefetchServiceEnabledBrowserTest.NoServePost
- SearchPrefetchServiceEnabledBrowserTest.NoServeReload

In addition to it, this CL adds the first party context check in the
storage key to inject the service worker. This will prevent synthetic
response being dispatched from the third-party context. This will fix
the following test in `LensOverlayControllerBrowserTest`.

- SidePanel_SearchURLClickWithTextDirective

`kNotEligibleByReload` is not recorded anymore in
`ServiceWorker.SyntheticResponse.Eligibility` since we don't inject the
service worker when reloading. UMA and related code will be cleaned up
in a follow up.


Bug: 352578800
Change-Id: Ib7671d251cf1fdde8c6c9c819da23f9a71acd12e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6893469
Reviewed-by: Keita Suzuki <suzukikeita@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1508982}